backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.0.post1
motor>=3.3.2
beanie>=1.25.0,<2.0.0
pydantic>=2.6.1
email-validator>=2.2.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-multipart>=0.0.9
redis[hiredis]>=5.0.1
boto3>=1.34.0
python-magic>=0.4.27; platform_system != "Windows"
python-magic-bin>=0.4.14; platform_system == "Windows"
openai>=1.14.0
pdfplumber>=0.11.0
pydub>=0.25.1
audioop-lts>=0.2.1; python_version >= "3.13"
tenacity>=8.2.3
langchain>=0.1.13
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.0.29
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
tiktoken>=0.6.0
slowapi>=0.1.9
